Main page content change
Changed the main page index content to reflect this being for a user instead of a project.

Changed css layout with more of a modern look with white background and bootstrap primary blue color.

In addition changed the text to Arial/Helvetica font

Added on medium (768px) and larger screens the font size increase to 18 px.

Added CSS custom properties (variables) for background, main, and text colors.